+++ filesystem/linux_support.cpp
@@ -102,7 +102,7 @@ HANDLE FindFirstFile( const char *fileName, FIND_DATA *dat)
 {
 	char nameStore[PATH_MAX];
 	char *dir=NULL;
-	intp n,iret=-1;
+	int n,iret=-1;
 	
 	Q_strncpy(nameStore,fileName, sizeof( nameStore ) );
