+++ game/server/ai_behavior_follow.h
@@ -100,7 +100,7 @@ struct AI_FollowGroup_t;
 struct AI_FollowManagerInfoHandle_t
 {
 	AI_FollowGroup_t *m_pGroup;
-	intp m_hFollower;
+	int m_hFollower;
 };
 
 //-------------------------------------
