+++ game/server/ai_hint.cpp
@@ -729,7 +729,7 @@ CAI_Hint *CAI_HintManager::GetFirstHint( AIHintIter_t *pIter )
 { 
 	if ( !gm_AllHints.Count() )
 	{
-		*pIter = (AIHintIter_t)(intp)gm_AllHints.InvalidIndex();
+		*pIter = (AIHintIter_t)gm_AllHints.InvalidIndex();
 		return NULL;
 	}
 	*pIter = (AIHintIter_t)0; 
@@ -741,12 +741,12 @@ CAI_Hint *CAI_HintManager::GetFirstHint( AIHintIter_t *pIter )
 //-----------------------------------------------------------------------------
 CAI_Hint *CAI_HintManager::GetNextHint(  AIHintIter_t *pIter )
 {
-	if ( (intp)*pIter != gm_AllHints.InvalidIndex() )
+	if ( (int)*pIter != gm_AllHints.InvalidIndex() )
 	{
-		intp i = ( (intp)*pIter ) + 1;
+		int i = ( (int)*pIter ) + 1;
 		if ( gm_AllHints.Count() <= i )
 		{
-			*pIter = (AIHintIter_t)(intp)gm_AllHints.InvalidIndex();
+			*pIter = (AIHintIter_t)gm_AllHints.InvalidIndex();
 			return NULL;
 		}
 		*pIter = (AIHintIter_t)i;
