+++ game/server/ai_navigator.cpp
@@ -1222,16 +1222,16 @@ float CAI_Navigator::GetPathTimeToGoal()
 AI_PathNode_t CAI_Navigator::GetNearestNode()
 {
 #ifdef WIN32
-	COMPILE_TIME_ASSERT( (intp)AIN_NO_NODE == NO_NODE );
+	COMPILE_TIME_ASSERT( (int)AIN_NO_NODE == NO_NODE );
 #endif
-	return (AI_PathNode_t)(intp)( GetPathfinder()->NearestNodeToNPC() );
+	return (AI_PathNode_t)( GetPathfinder()->NearestNodeToNPC() );
 }
 
 //-----------------------------------------------------------------------------
 
 Vector CAI_Navigator::GetNodePos( AI_PathNode_t node )
 {
-	return GetNetwork()->GetNode((intp)node)->GetPosition(GetHullType());
+	return GetNetwork()->GetNode((int)node)->GetPosition(GetHullType());
 }
 
 //-----------------------------------------------------------------------------
