+++ ivp/ivp_compact_builder/ivp_compact_recursive.cxx
@@ -63,7 +63,7 @@ void IVP_Compact_Recursive::build_convex_hull(){
 
 void IVP_Compact_Recursive::set_rekursive_convex_hull(){
   IVP_Hash point_hash( 1024, sizeof(IVP_U_Float_Point),(void *)-1);
-  intp n_points_in_hash = 0;
+  int n_points_in_hash = 0;
   
   struct Triangle_Key {
     int point_index[3];
@@ -103,7 +103,7 @@ void IVP_Compact_Recursive::set_rekursive_convex_hull(){
       int pi[3];
       // convert points to unique point nums
       for (int i = 0; i<3;i++){
-	intp ind = (intp)point_hash.find( (char *)p[i]);
+	int ind = (int)point_hash.find( (char *)p[i]);
 	if (ind<0){
 	  point_hash.add( (char *)p[i], (void *)n_points_in_hash);
 	  ind = n_points_in_hash++;
@@ -140,7 +140,7 @@ void IVP_Compact_Recursive::set_rekursive_convex_hull(){
       int pi[3];
       // convert points to unique point nums
       for (int i = 0; i<3;i++){
-    intp ind = (intp)point_hash.find( (char *)p[i]);
+	int ind = (int)point_hash.find( (char *)p[i]);
 	pi[i] = ind;
       }
       // find triangle in hash
