+++ public/phyfile.h
@@ -11,14 +11,13 @@
 
 #include "datamap.h"
 
-
 typedef struct phyheader_s
 {
 	DECLARE_BYTESWAP_DATADESC();
 	int		size;
 	int		id;
 	int		solidCount;
-	int		checkSum;	// checksum of source .mdl file
+	long	checkSum;	// checksum of source .mdl file
 } phyheader_t;
 
 #endif // PHYFILE_H
