+++ tier2/soundutils.cpp
@@ -29,22 +29,22 @@ class CFSIOReadBinary : public IFileReadBinary
 {
 public:
 	// inherited from IFileReadBinary
-	virtual intp open( const char *pFileName );
-	virtual int read( void *pOutput, int size, intp file );
-	virtual void seek( intp file, int pos );
-	virtual unsigned int tell( intp file );
-	virtual unsigned int size( intp file );
-	virtual void close( intp file );
+	virtual int open( const char *pFileName );
+	virtual int read( void *pOutput, int size, int file );
+	virtual void seek( int file, int pos );
+	virtual unsigned int tell( int file );
+	virtual unsigned int size( int file );
+	virtual void close( int file );
 };
 
 class CFSIOWriteBinary : public IFileWriteBinary
 {
 public:
-	virtual intp create( const char *pFileName );
-	virtual int write( void *pData, int size, intp file );
-	virtual void close( intp file );
-	virtual void seek( intp file, int pos );
-	virtual unsigned int tell( intp file );
+	virtual int create( const char *pFileName );
+	virtual int write( void *pData, int size, int file );
+	virtual void close( int file );
+	virtual void seek( int file, int pos );
+	virtual unsigned int tell( int file );
 };
 
 
@@ -61,12 +61,12 @@ IFileWriteBinary *g_pFSIOWriteBinary = &s_FSIoOut;
 //-----------------------------------------------------------------------------
 // RIFF reader that use the file system
 //-----------------------------------------------------------------------------
-intp CFSIOReadBinary::open( const char *pFileName )
+int CFSIOReadBinary::open( const char *pFileName )
 {
-	return (intp)g_pFullFileSystem->Open( pFileName, "rb" );
+	return (int)g_pFullFileSystem->Open( pFileName, "rb" );
 }
 
-int CFSIOReadBinary::read( void *pOutput, int size, intp file )
+int CFSIOReadBinary::read( void *pOutput, int size, int file )
 {
 	if ( !file )
 		return 0;
@@ -74,7 +74,7 @@ int CFSIOReadBinary::read( void *pOutput, int size, intp file )
 	return g_pFullFileSystem->Read( pOutput, size, (FileHandle_t)file );
 }
 
-void CFSIOReadBinary::seek( intp file, int pos )
+void CFSIOReadBinary::seek( int file, int pos )
 {
 	if ( !file )
 		return;
@@ -82,7 +82,7 @@ void CFSIOReadBinary::seek( intp file, int pos )
 	g_pFullFileSystem->Seek( (FileHandle_t)file, pos, FILESYSTEM_SEEK_HEAD );
 }
 
-unsigned int CFSIOReadBinary::tell( intp file )
+unsigned int CFSIOReadBinary::tell( int file )
 {
 	if ( !file )
 		return 0;
@@ -90,7 +90,7 @@ unsigned int CFSIOReadBinary::tell( intp file )
 	return g_pFullFileSystem->Tell( (FileHandle_t)file );
 }
 
-unsigned int CFSIOReadBinary::size( intp file )
+unsigned int CFSIOReadBinary::size( int file )
 {
 	if ( !file )
 		return 0;
@@ -98,7 +98,7 @@ unsigned int CFSIOReadBinary::size( intp file )
 	return g_pFullFileSystem->Size( (FileHandle_t)file );
 }
 
-void CFSIOReadBinary::close( intp file )
+void CFSIOReadBinary::close( int file )
 {
 	if ( !file )
 		return;
@@ -110,28 +110,28 @@ void CFSIOReadBinary::close( intp file )
 //-----------------------------------------------------------------------------
 // RIFF writer that use the file system
 //-----------------------------------------------------------------------------
-intp CFSIOWriteBinary::create( const char *pFileName )
+int CFSIOWriteBinary::create( const char *pFileName )
 {
 	g_pFullFileSystem->SetFileWritable( pFileName, true );
-	return (intp)g_pFullFileSystem->Open( pFileName, "wb" );
+	return (int)g_pFullFileSystem->Open( pFileName, "wb" );
 }
 
-int CFSIOWriteBinary::write( void *pData, int size, intp file )
+int CFSIOWriteBinary::write( void *pData, int size, int file )
 {
 	return g_pFullFileSystem->Write( pData, size, (FileHandle_t)file );
 }
 
-void CFSIOWriteBinary::close( intp file )
+void CFSIOWriteBinary::close( int file )
 {
 	g_pFullFileSystem->Close( (FileHandle_t)file );
 }
 
-void CFSIOWriteBinary::seek( intp file, int pos )
+void CFSIOWriteBinary::seek( int file, int pos )
 {
 	g_pFullFileSystem->Seek( (FileHandle_t)file, pos, FILESYSTEM_SEEK_HEAD );
 }
 
-unsigned int CFSIOWriteBinary::tell( intp file )
+unsigned int CFSIOWriteBinary::tell( int file )
 {
 	return g_pFullFileSystem->Tell( (FileHandle_t)file );
 }
